refactor(validation): arrange modules related to http-server into validation/http directory
