+++ src/routers/chats-router.ts
@@ -1,6 +1,6 @@
 import express from 'express';
 import { chatsController } from '../controllers/chats-controller';
-import { validateId } from '../validation/validator';
+import { validateId } from '../validation/http/validator';
 
 export const chatsRouter = express.Router();
